Platform/Ampere: Add RamBlockIoDxe module
This module produces BlockIoProtocol on a fixed memory address. It
allows a virtual harddisk device (VHP) which is copied to a fixed memory
address to be mounted and recognized by UEFI.

For example, create a 32MB VHP with FAT32 format.
  dd if=/dev/zero of=vhd.img bs=1M count=32
  mkfs.fat -F 32 vhd.img

Then copy vhd.img to the fixed memory address specified in RamBlockIoDxe
before UEFI boot. Make sure the address is untouched by any codes.

Note that this driver is currently not used at this time. It should be
added to platform build for use.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>
Signed-off-by: Phong Vo <pvo@amperecomputing.com>