Platform/Ampere: Add LinuxBoot image
LinuxBoot is a firmware that replaces specific firmware functionality
like the UEFI DXE phase with a Linux kernel and runtime.

This image is used for replacing the UEFI Shell Application in case we
want to execute the LinuxBoot at end of the DXE phase.

Also add the kernel config file that is used to generate the kernel image.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>
Signed-off-by: Vanshidhar Konda <vkonda@amperecomputing.com>
Signed-off-by: Phong Vo <pvo@amperecomputing.com>