JadePkg: Support LinuxBoot DSC/FDF build for Jade platform
This adds new JadeLinuxboot/Ac01LinuxbootPkg DSC and JadeLinuxboot FDF
file to support the LinuxBoot build. The LinuxBoot image will be built
into the Tianocore EDK2 image as replacing the UEFI Shell Application.
Some PEI/DXE drivers are not added into the platform DSC/FDF as it is
not required for booting to LinuxBoot Shell (u-root).

As some recent changes from ArmPkg/PlatformBootManagerLib, the change
hide UEFI Shell as a regular boot option. The UEFI Shell is not invoke
automatically when no active boot options could be started. Instead, it
drop into UiApp (Setup Screen) and we have to enter UEFI Shell manually
to load the LinuxBoot.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>
Signed-off-by: Phong Vo <pvo@amperecomputing.com>