Silicon/Ampere: Implement PlatformBootManagerLib for LinuxBoot
This PlatformBootManagerLib is based on the ArmPkg version. It is
modified to support to boot LinuxBoot Shell immediately from an active
boot option, without entering UiApp Setup Screen.

The size of LinuxBoot FD is 0x00B00000. This is also the  maximum size
of UEFI Partition in the SPI-Nor Layout that is defined in the Altra
Interface Firmware Spec. It leads to the size of the FD overflowed from
UEFI partition when the Trusted Firmware (ATF) TBB is enabled, due to
the size of the FD is increased by an additional BL33 certificate.
Therefore, the FD size will be reduced to 0x00AC0000 to make the
LinuxBoot can boot properly both TBB=1 and TBB=0.  To keep the MAIN FV
size is unchanged, the PEI FV is reduced to 0x00050000 to keep the
MAIN FV big enough.  The BlockSize in the FD and FV is also updated to
reflect the correct block size 64KB. It's required by UEFI Runtime DXE.

In the future, this implementation is used for customizing the BDS phase
for LinuxBoot specific.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>
Signed-off-by: Phong Vo <pvo@amperecomputing.com>