AmpereAltraPkg: Combine Ampere CPU libraries
The AmpereCpuLib and CpuCommonLib have the same functionality that
provide APIs related to Ampere CPU control/info functions. It should be
combined into one library for being simple to use and manageable. So,
this patch merges the CpuCommonLib into AmpereCpulib library and renames
the Library/PlatformInfo.h to PlatformInfoHob.h for platform info Hob
structures. Also, this removes the unused PlatformInfoLib.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>