Platform/Ampere: Add FailSafe and WDT support
The FailSafeDxe driver supports to revert the system's configuration to
last good values if the system fails to boot up multiple times. It also
implements the Watchdog Timer Architectural Protocol to reset the system
if it hangs.

By default, when system starts, it configures the secure watchdog timer
with a default value of 5 minutes. If the system boots up cleanly to a
considered good stage, the counter is cleared as it indicates FailSafe
monitor (ATF) that the system boots up successfully. If the system boots
up with failure, it will be restarted right after the timer expires.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>