Ampere: ACPI: Remove check for installing ACPI multiple times
Checking whether an ACPI table is available or not before installing
seems to be workaround when the ReadyToBoot event callback is called
multiple times. This issue was already resolved by closing the event
at the callback function to prevent it to be signalled again. This
patch removes the check as they are redundant now.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>