Silicon/Ampere: Fix MemoryAddressSize in IORT Root complex Nodes
The Root complex nodes in the IORT should set the MemoryAddressSize
entry in the template struct used to initialize the Root complex nodes.

Failing to do so leads to corrupted DMA masks on newer Linux kernels
that attempt to actually use the entries without properly guarding
against it being unset.

An upstream kernel fix that will handle unset entries is in progress.

Signed-off-by: Moritz Fischer <mdf@kernel.org>
Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>