JadePkg: ACPI EINJ: Fix GET_BUSY_STATUS field
Some operating systems do not follow the EINJ GET_BUSY_STATUS values
that are defined in the ACPI spec. Instead they rely on the EINJ
GET_BUSY_STATUS read register value to indicate when the platform
is busy with an EINJ request. Update the GET_BUSY_STATUS read
register value to 1 so that both OS implementations will work as
intended.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>