AmpereSiliconPkg: Adjust IPMI SSIF response retry timeout
Currently, the maximum waiting time-out for a response is about 300ms.
However, BMC might take a longer time to prepare the response data for
some IPMI commands. This change increases the maximum waiting time-out
same as Linux IPMI SSIF driver with number of retries 250 and the
time-out 20ms for each.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>