AmpereAltraPkg/RasConfigDxe: Isolate RAS Configuration from APEI
The ACPI Platform Error Interfaces (APEI) provides a means for a
platform to convey error information to OSPM. The APEI consists of
separate tables such as Boot Error Record Table (BERT), Hardware Error
Source Table (HEST), Error Injection Table (EINJ),...

The RAS Configuration screen provides options to control the RAS
features in the system firmware like ATF and SMpro/PMpro.

So, the RAS Configuration should be isolated from APEI support. With
APEI disabled in the ACPI Configuration, it should uninstall the APEI
tables, but not disable the RAS Configuration. It makes sure that the
RAS features are still working in SMpro/PMpro and reporting to BMC.

Signed-off-by: Nhi Pham <npham@amperecomputing.com>