AmpereAltraPkg: Add FailSafe and WDT support
The FailSafeDxe is a driver for the FailSafe feature which reverts the
system's configuration to known good values if the system fails to boot
up multiple times. Also, this driver implements the Watchdog Timer
Architectural Protocol to reset the system if it hangs, which is
implemented by the MdeModulePkg/Universal/WatchdogTimerDxe module. So,
the WDT is now used exclusively by the FailSafeDxe.

By default, when system starts, it configures the secure watchdog timer
with a default value of 5 minutes. If the system boots up cleanly to the
considered good stage, the counter is cleared as it indicates FailSafe
monitor (ATF) that has booted up successfully. If the timer expires, it
is considered a failed boot and the system is rebooted.

Cc: Thang Nguyen <thang@os.amperecomputing.com>
Cc: Chuong Tran <chuong@os.amperecomputing.com>
Cc: Phong Vo <phong@os.amperecomputing.com>
Cc: Leif Lindholm <leif@nuviainc.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Nate DeSimone <nathaniel.l.desimone@intel.com>

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>
Reviewed-by: Leif Lindholm <leif@nuviainc.com>