AmpereAltraPkg: IpmiBoot: Change to restore the BootOrder on next boot
Currently, if persistent option is not set, the code will try to restore
the BootOrder to previous value right after forced option was successfully
called. The point when the restore process happens depends on the option
types, such as: ExitBootService for the disk option or ReadyToBoot for PXE.
Unfortunately, on first boot of a system with TPM, CentOS Shim might trigger
system reset before ExitBootService and prevent the BootOrder restore from
happening. To solve these kinds of issues and simplify the restore process,
refactor the code to restore the BootOrder on next boot instead of rely on
various system event. This change also fixes the boot to UiApp issue when
NVRAM was cleared by limiting eligible cases.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>