JadePkg: Add driver to produce SMBIOS Type 42
Define SMBIOS handle for Redfish communication over USB CDC ethernet
interface. The info of this handle will be consumed by the RedfishPkg
drivers for their configuration.
Current RedfishHostInterfaceDxe already supports this feature but the
problem is that it requires a definite MAC address of the client NIC
which is not available at the early boot stage.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>