AmpereAltraPkg: Support DDR CE Threshold Window Filter
This change is to limit the number of DDR CEs which will be reported
to the BMC and OS to avoid the CE storm. The default DDR CE will be
hide in the RAS menu config if this feature is enabled.

The control to this feature can be done through the following board
setting entries:
- NV_SI_RO_BOARD_RAS_DDR_CE_WINDOW: CE Windows (minute unit)
  TH1 and TH2 counters reset when period has elapsed
  Default value is 0 -- equals disabled
- NV_SI_RO_BOARD_RAS_DDR_CE_TH1: CE threshold fine (TH1)
  Low threshold to report CE to BMC and OS
  Default value is 500
- NV_SI_RO_BOARD_RAS_DDR_CE_TH2: CE threshold coarse (TH2)
  High threshold to stop reporting until CE Window has elapsed.
  Default value is 5000

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>