AmperePlatformPkg/FailSafeLib: Notify ATF on success boot
ATF must save the working NVPARAM values after receiving the
notification from the UEFI. Currently, UEFI refuses to do that
if the checksum of the FailSafe context is invalid. Since this
region is untouched unless having any issues with all previous
boot cycles, It will raise a problem that the latest NVPARAM does
not backup, and ATF still uses the default values for future
NVPARAM restore.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>