AmpereSiliconPkg: Fix failed to get Redfish account via IPMI
Currently, OBMC will return 1-byte data with the CRC error in response
to the host request if the data is not ready. The host must be aware of
this and continuously poll the BMC until receiving valid data. However,
at present, errors arising while reading the data are being ignored. By
returning the error code immediately, this patch will help to address
this problem.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>