Ampere: Change to use MmCommunicationDxe from ArmPkg
It's unnecessary to duplicate a ready-to-use module while we can change
our code to adapt it. This change is to remove the Ampere version of
MmCommunicationDxe and turn to use the one from ArmPkg. It also changed
from MmCommunicationProtocol to MmCommunication2Protocol with a bit
update in the function parameters.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>