AmpereAltraPkg/Tcg2Pei: Fix TPM event log corrupted
This patch fixes two issues:
* The Pre-UEFI firmware components only support one algorithm at a time.
For example, if the SHA256 is selected to support then the hash of SHA1
is empty. But UEFI expects both SHA1 and SHA256. To solve this issue,
the ComplementPreUefiDigest function is implemented, which does
measurement of the measured digest. In case the pre-UEFI firmware only
supports the SHA1, the SHA1 digest will be used as a hash data for
SHA256 measurement.
* Per the design, the "Post SCP TPM Extend" (PSTE) is already extended
in ATF. Therefore, the extend of PSTE event must be skipped and just log
the event.

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>