AmpereAltraPkg: TPM2: Enable PP operation in TCG2 configuration screen
This patch cleans up the configuration screen for TPM2 and adds the TPM
control like PPI Request and PCR selection to be active.

Currently, the PPI Requestion option in the TCG2 Configuration screen
only supports the PP Clear command because of the limitation of the
TPM firmware interface. Follow the below steps to verify this support:
1/ Press <Enter> at the TPM2 Operation entry
2/ Select the "TPM2 ClearControl(NO) + Clear" option for the request
3/ The following message is displayed in the ATF console:
  "NOTICE:  TPM: PPI: Request for operation code 5 is registered."
  The following message is displayed in the SCP console:
  "TPM:PPI: NS-DB 2 interrupt arrived"
4/ Reset the system
5/ The TPM operation gets reflected and corresponding successful
operation message is displayed in the ATF console as follows:
  "NOTICE:  TPM: PPI: Request found.
  NOTICE:  TPM: PPI: ClearControl(No) + Clear() Successful
  NOTICE:  TPM: Discovery completes"

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>