JadePkg: Fix booting to UiApp when NVRAM cleared issue
Add new variable to indicate the force boot to UiApp instead of checking
the first boot option as it will raise an issue when booting with NVRAM
cleared. This commit also change to process the IPMI boot option at the
driver entrypoint instead of waiting until End Of Dxe event since the
BootOrder in both case are the same.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>