JadePkg/AcpiTables: Update for new DSU PMU spec
To follow new DSU PMU specs, the following changes has been made:
- Add a cluster node for each CPM.
- Add DSU node to each CPM.

Because LPI requires _LPI node for all the level of processor hierarchy.
DSU adds _LPI to the cluster level. To do that,
- Add _LPI S-WFI state for each core
- Add dummy _LPI state for each cluster. Ampere Altra doesn't support
cluster idle state, this node is just a middle state which does nothing
in term of power of system.

Signed-off-by: Hoan Tran <hoan.tran@amperecomputing.com>
Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>