JadePkg/AcpiTables: Remove LPI-1 Standby ATF state
With the current design, the system has 3 LPI states
 LPI0: CPU Standby (WFI) in OSPM/EL1
 LPI1: CPU Standby (WFI) in ATF/EL3
 LPI2: System Standby and all core goes into WFI at ATF/EL3

The LPI1 is the same with LPI0 which only goes into WFI. It does
not save much power but it costs CPU cycle to go into ATF. Because
of that, this change eliminate LPI1. The system only has 2 LPI states
 LPI0: CPU Stanby (WFI) in OSPM/EL1
 LPI1: System standby and all core goes into WFI in ATF/EL3. CPUPLL
frequency is reduced.

Signed-off-by: Hoan Tran <hoan.tran@amperecomputing.com>
Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>