AmpereAltraPkg/IpmiBootDxe: Fix failed to set bootdev when NVRAM cleared
Current design does not allow to handle the IPMI set bootdev when NVRAM
cleared but delay to next boot cycle. With the target to process bootdev
in the current boot cycle, this change moves the IPMI bootdev handler to
After Console event as all boot options are expected to be available at
this time. UiApp forcing approach is also changed from setting BootNext
to OsIndication to adapt with above change.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>