JadePkg: Add information of L3 cache in SMBIOS when SLC enabled
The System Level Cache (SLC) information will be added to the SMBIOS
Type 7 as L3 cache when the SLC as L3$ option is enabled in the CPU
Configuration screen.

With the SLC as L3$ option enabled, the SMBIOS Type 7 will contain L3
cache information and the SMBIOS Type 4 will contain its handle.

Signed-off-by: Tinh Nguyen <tinhn@amperecomputing.com>