AmpereAltraPkg: Fix default Processor Asset Tag information (SMBIOS type 4)
In FWTS, there's a test case checking whether Asset Tag field of SMBIOS
type 4 is updated or not. If yes, the information must be different with
default information (Not Specified, Unknown, ...). But EDK2 is using
"Not Specified" as a default information for this field so it fails at
certain test case in FWTS. This update helps to avoid that issue.

Signed-off-by: Minh Nguyen <minhn@amperecomputing.com>